# docs/swagger.yaml
basePath: /
definitions:
  models.Menu:
    properties:
      category:
        type: string
      created_at:
        type: string
      end_date:
        type: string
      food_id:
        type: string
      id:
        type: string
      name:
        type: string
      start_date:
        type: string
      updated_at:
        type: string
    required:
    - category
    - name
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8080
info:
  contact:
    email: abhiroop.santra@gmail.com
    name: Abhiroop Santra
    url: https://www.abhiroopsantra.dev/
  description: This is the Backend API for Gastrono Go.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Gastrono Go Backend API
  version: "1.0"
paths:
  /menus:
    get:
      consumes:
      - application/json
      description: Get All Menus
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Menu'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get All Menus
      tags:
      - menu
  /menus/{menu_id}:
    get:
      consumes:
      - application/json
      description: Get All Menus
      parameters:
      - description: Menu ID
        in: path
        name: menu_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Menu'
            type: array
      summary: Get All Menus
      tags:
      - menu
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"